+++ recipes/default.rb
@@ -1,47 +1,47 @@
 #
-# Cookbook Name:: git-server
+# Cookbook Name:: chef-git-server
 # Recipe:: default
 #
 #
 
 # Create git user on server
-user node['git-server']['user'] do
+user node['chef-git-server']['user'] do
   supports :manage_home => true
-  comment node['git-server']['user_comment']
-  home node['git-server']['home']
-  shell node['git-server']['shell']
+  comment node['chef-git-server']['user_comment']
+  home node['chef-git-server']['home']
+  shell node['chef-git-server']['shell']
 end
 
-directory File.join(node['git-server']['home'], ".ssh") do
-  user node['git-server']['user']
-  group node['git-server']['group']
+directory File.join(node['chef-git-server']['home'], ".ssh") do
+  user node['chef-git-server']['user']
+  group node['chef-git-server']['group']
   mode "700"
 end
 
 # Pulls all SSH Keys out of users databag and adds to the git user
 # authorized_keys.  See users cookbook for details"
-users = data_bag(node['git-server']['user_data_bag'])
+users = data_bag(node['chef-git-server']['user_data_bag'])
 ssh_keys = ''
 users.each do |username|
-  user = data_bag_item(node['git-server']['user_data_bag'], node['git-server']['username_data_bag'])
-  Array(user[node['git-server']['ssh_keys_data_bag']]).each do |ssh_key|
+  user = data_bag_item(node['chef-git-server']['user_data_bag'], node['chef-git-server']['username_data_bag'])
+  Array(user[node['chef-git-server']['ssh_keys_data_bag']]).each do |ssh_key|
     ssh_keys << ssh_key + "\n"
   end
 end
 
-file File.join(node['git-server']['home'], File.join('.ssh', 'authorized_keys')) do
-  owner node['git-server']['user']
-  group node['git-server']['group']
+file File.join(node['chef-git-server']['home'], File.join('.ssh', 'authorized_keys')) do
+  owner node['chef-git-server']['user']
+  group node['chef-git-server']['group']
   mode "600"
   content ssh_keys
 end
 
 # Setup repositories defined as node attributes
-node['git-server']['repositories'].each do |repository_name|
+node['chef-git-server']['repositories'].each do |repository_name|
   execute "git init --bare #{repository_name}.git" do
-    user node['git-server']['user']
-    group node['git-server']['group']
-    cwd node['git-server']['home']
-    creates File.join(node['git-server']['home'], "#{repository_name}.git")
+    user node['chef-git-server']['user']
+    group node['chef-git-server']['group']
+    cwd node['chef-git-server']['home']
+    creates File.join(node['chef-git-server']['home'], "#{repository_name}.git")
   end
 end